requirements.txt
discord-py==2.3.2
flask==3.0.0
python-dotenv
Pillow
